@@ -1,16 +1,16 @@
-name: flutter_mongo_stitch_platform_interface
+name: flutter_mongo_stitch_platform_interface_besaved
 description: A new Flutter package project.
-version: 1.2.2
-homepage: https://github.com/kfiross/flutter_mongo_stitch_platform_interface
+version: 1.2.4
+homepage: https://github.com/BesavedTech/flutter_mongo_stitch_platform_interface_bst
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
 
-  plugin_platform_interface: ^1.0.3
+  plugin_platform_interface: ^2.0.0
 
 dev_dependencies:
   flutter_test:
